refactor: simplify build and packaging of button-group component
