+++ packages/components/button-group/css/index.js
@@ -1,9 +1,6 @@
-
-export { BsButtonGroup } from './bs-button-group.js';
 export { BsButtonGroupHorizontalCss } from './bs-button-group-horizontal.css.js';
 export { BsButtonGroupLargeCss } from './bs-button-group-large.css.js';
 export { BsButtonGroupSmallCss } from './bs-button-group-small.css.js';
 export { BsButtonGroupToggleCss } from './bs-button-group-toggle.css.js';
 export { BsButtonGroupVerticalCss } from './bs-button-group-vertical.css.js';
-export { BsButtonToolbar } from './bs-button-toolbar.js';
 export { BsButtonToolbarCss } from './bs-button-toolbar.css.js';
